Editing add record function

+++ clinicManagementSystem/database/migrations/2023_06_11_071204_create_bills_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('patient_id');
             $table->unsignedBigInteger('record_id');
-            $table->decimal('total_bill',6,3);
+            $table->integer('total_bill');
             $table->timestamps();
             $table->foreign('patient_id')->references('id')->on('patients')->onDelete('cascade');
             $table->foreign('record_id')->references('id')->on('records')->onDelete('cascade');
